Add support for semantic C++ types
Changing FFIStringCpp to something that is not considered a POD type by the type
system revealed a subtle, but previously effectless issue. The custom _by
comparison constructs treated the types as the original FFI types and not as the
derived C++ types. This did not have any effect because all derived types had
the same POD semantics, and the comparison function was provided by the Rust
side. But with FFIStringCpp changing to a non-trivially copyable move only type,
this issue was revealed. The fix is to provide the comparison constructs with
the derived type and to take a comparison function-pointer as generic template
parameter to decouple T and the parameter type of the function-pointer. In
addition the Rust side now understands FFIString in a moved from state and can
safely destroy such values.